+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/AthenaApiRepository.kt
@@ -14,6 +14,7 @@ import software.amazon.awssdk.services.athena.model.StopQueryExecutionRequest
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Dataset
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Datasource
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.FilterType.Date
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.MultiphaseQuery
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ReportFilter
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ReportSummary
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementCancellationResponse
@@ -22,11 +23,17 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshif
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.DprAuthAwareAuthenticationToken
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.TableIdGenerator
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.model.Prompt
+import java.util.Base64
 
 const val QUERY_STARTED = "STARTED"
 const val QUERY_FINISHED = "FINISHED"
 const val QUERY_ABORTED = "ABORTED"
 const val QUERY_FAILED = "FAILED"
+const val QUERY_SUCCEEDED = "SUCCEEDED"
+const val QUERY_CANCELLED = "CANCELLED"
+const val QUERY_RUNNING = "RUNNING"
+const val QUERY_QUEUED = "QUEUED"
+const val QUERY_SUBMITTED = "SUBMITTED"
 
 @Service
 @Primary
@@ -56,6 +63,200 @@ class AthenaApiRepository(
     reportOrDashboardId: String,
     reportOrDashboardName: String,
     preGeneratedDatasetTableId: String?,
+    multiphaseQuery: List<MultiphaseQuery>?,
+  ): StatementExecutionResponse = multiphaseQuery?.takeIf { it.isNotEmpty() }?.let {
+    executeMultiphaseQuery(
+      productDefinitionId,
+      productDefinitionName,
+      reportOrDashboardId,
+      reportOrDashboardName,
+      userToken,
+      prompts,
+      reportFilter,
+      policyEngineResult,
+      sortColumn,
+      sortedAsc,
+      datasource,
+      it,
+    )
+  }
+    ?: executeSingleQuery(
+      productDefinitionId,
+      productDefinitionName,
+      reportOrDashboardId,
+      reportOrDashboardName,
+      userToken,
+      prompts,
+      query,
+      reportFilter,
+      policyEngineResult,
+      dynamicFilterFieldId,
+      sortColumn,
+      sortedAsc,
+      datasource,
+    )
+
+  private fun executeMultiphaseQuery(
+    productDefinitionId: String,
+    productDefinitionName: String,
+    reportOrDashboardId: String,
+    reportOrDashboardName: String,
+    userToken: DprAuthAwareAuthenticationToken?,
+    prompts: List<Prompt>?,
+    reportFilter: ReportFilter?,
+    policyEngineResult: String,
+    sortColumn: String?,
+    sortedAsc: Boolean,
+    datasource: Datasource,
+    multiphaseQuery: List<MultiphaseQuery>,
+  ): StatementExecutionResponse {
+    // consider executing one insert operation
+    val multiphaseQuerySortedByIndex = multiphaseQuery.sortedBy { it.index }
+    val firstTableId = tableIdGenerator.generateNewExternalTableId()
+    val firstQuery = """
+            /* $productDefinitionId $productDefinitionName $reportOrDashboardId $reportOrDashboardName */
+            CREATE TABLE AwsDataCatalog.reports.$firstTableId 
+            WITH (
+              format = 'PARQUET'
+            ) 
+            AS (
+            SELECT * FROM TABLE(system.query(query =>
+             '${(
+      listOf(buildContextQuery(userToken), buildPromptsQuery(prompts), buildDatasetQuery(multiphaseQuerySortedByIndex[0].query))
+        .joinToString(",") +
+        "\nSELECT * FROM $DATASET_"
+      ).replace("'", "''")}'
+             )) 
+            );
+    """.trimIndent()
+    log.debug("Database query at index ${multiphaseQuerySortedByIndex[0].index}: $firstQuery")
+    val firstQueryDatasource = multiphaseQuerySortedByIndex[0].datasource
+    val statementExecutionResponse = executeQueryAsync(firstQueryDatasource, firstTableId, firstQuery)
+    val jdbcTemplate = populateJdbcTemplate(firstQueryDatasource.name)
+    val insertStatement: String =
+      buildInsertStatement(
+        rootExecutionId = statementExecutionResponse.executionId,
+        currentExecutionId = statementExecutionResponse.executionId,
+        datasourceName = firstQueryDatasource.name,
+        datasourceCatalog = firstQueryDatasource.catalog,
+        datasourceDatabase = firstQueryDatasource.database,
+        index = multiphaseQuerySortedByIndex[0].index,
+        query = firstQuery,
+      )
+    log.debug("Inserting into admin table: {}", insertStatement)
+    jdbcTemplate.execute(insertStatement)
+    var previousTableId = firstTableId
+    List(
+      multiphaseQuerySortedByIndex
+        .drop(1)
+        .dropLast(1).size,
+    ) { i ->
+      val currentTableId = tableIdGenerator.generateNewExternalTableId()
+      val intermediateQuery = multiphaseQuerySortedByIndex[i + 1]
+      val intermediateQueryString = """
+          /* $productDefinitionId $productDefinitionName $reportOrDashboardId $reportOrDashboardName */
+              CREATE TABLE AwsDataCatalog.reports.$currentTableId
+              WITH (
+                format = 'PARQUET'
+              ) 
+              AS (
+              ${(
+        listOf(buildContextQuery(userToken, false), buildPromptsQuery(prompts, false), buildDatasetQuery(intermediateQuery.query))
+          .joinToString(",") +
+          "\nSELECT * FROM $DATASET_"
+        ) .replace("\${tableId}", previousTableId)}
+              )
+      """.trimIndent()
+      log.debug("Intermediate query at index ${i + 1}: {}", intermediateQueryString)
+      val insertQuery = buildInsertStatement(
+        rootExecutionId = statementExecutionResponse.executionId,
+        datasourceName = intermediateQuery.datasource.name,
+        datasourceCatalog = intermediateQuery.datasource.catalog,
+        datasourceDatabase = intermediateQuery.datasource.database,
+        index = intermediateQuery.index,
+        query = intermediateQueryString,
+      )
+      log.debug("Inserting into admin table: {}", insertQuery)
+      jdbcTemplate.execute(insertQuery)
+      previousTableId = currentTableId
+    }
+    val lastTableId = tableIdGenerator.generateNewExternalTableId()
+    val lastQuery = """
+      /* $productDefinitionId $productDefinitionName $reportOrDashboardId $reportOrDashboardName */
+              CREATE TABLE AwsDataCatalog.reports.$lastTableId
+              WITH (
+                format = 'PARQUET'
+              ) 
+              AS (
+               ${buildFinalQuery(
+      buildContextQuery(userToken, false),
+      buildPromptsQuery(prompts, false),
+      buildDatasetQuery(multiphaseQuerySortedByIndex.last().query),
+      buildReportQuery(reportFilter),
+      buildPolicyQuery(policyEngineResult, determinePreviousCteName(reportFilter)),
+      "$FILTER_ AS (SELECT * FROM $POLICY_ WHERE $TRUE_WHERE_CLAUSE)",
+      buildFinalStageQuery(sortColumn = sortColumn, sortedAsc = sortedAsc),
+    )
+      .replace("\${tableId}", previousTableId)}
+              )
+    """.trimIndent()
+    log.debug("Last multiphase query: {}", lastQuery)
+    val lastInsertStatement = buildInsertStatement(
+      rootExecutionId = statementExecutionResponse.executionId,
+      datasourceName = multiphaseQuerySortedByIndex.last().datasource.name,
+      datasourceCatalog = multiphaseQuerySortedByIndex.last().datasource.catalog,
+      datasourceDatabase = multiphaseQuerySortedByIndex.last().datasource.database,
+      index = multiphaseQuerySortedByIndex.last().index,
+      query = lastQuery,
+    )
+    log.debug("Inserting into admin table: {}", lastInsertStatement)
+    jdbcTemplate.execute(lastInsertStatement)
+    return StatementExecutionResponse(lastTableId, statementExecutionResponse.executionId)
+  }
+
+  private fun buildInsertStatement(
+    rootExecutionId: String,
+    currentExecutionId: String? = null,
+    datasourceName: String,
+    datasourceCatalog: String? = null,
+    datasourceDatabase: String? = null,
+    index: Int,
+    query: String,
+  ) = """insert into 
+          admin.execution_manager (
+          root_execution_id,
+          ${currentExecutionId?.let { "current_execution_id,"} ?: ""}
+          datasource,
+          ${datasourceCatalog?.let { "catalog,"} ?: "" }
+          ${datasourceDatabase?.let { "database," } ?: "" }
+          index,
+          query
+          )
+          values (
+            '$rootExecutionId',
+            ${currentExecutionId?.let { "'$it'," } ?: ""}
+            '$datasourceName',
+            ${datasourceCatalog?.let { "'$it'," } ?: ""}
+            ${datasourceDatabase?.let { "'$it'," } ?: ""}
+            $index,
+            '${Base64.getEncoder().encodeToString(query.toByteArray())}'
+          )
+  """.trimMargin()
+
+  private fun executeSingleQuery(
+    productDefinitionId: String,
+    productDefinitionName: String,
+    reportOrDashboardId: String,
+    reportOrDashboardName: String,
+    userToken: DprAuthAwareAuthenticationToken?,
+    prompts: List<Prompt>?,
+    query: String,
+    reportFilter: ReportFilter?,
+    policyEngineResult: String,
+    dynamicFilterFieldId: Set<String>?,
+    sortColumn: String?,
+    sortedAsc: Boolean,
+    datasource: Datasource,
   ): StatementExecutionResponse {
     val tableId = tableIdGenerator.generateNewExternalTableId()
     val finalQuery = """
@@ -136,12 +337,12 @@ class AthenaApiRepository(
 
   override fun buildDatasetQuery(query: String) = if (query.contains("$DATASET_ AS", ignoreCase = true)) query else """$DATASET_ AS ($query)"""
 
-  private fun buildPromptsQuery(prompts: List<Prompt>?): String {
+  private fun buildPromptsQuery(prompts: List<Prompt>?, useDualTable: Boolean = true): String {
     if (prompts.isNullOrEmpty()) {
-      return "$PROMPT AS (SELECT '' FROM DUAL)"
+      return "$PROMPT AS (SELECT '' ${if (useDualTable) "FROM DUAL" else ""})"
     }
     val promptsCte = prompts.joinToString(", ") { prompt -> buildPromptsQueryBasedOnType(prompt) }
-    return "$PROMPT AS (SELECT $promptsCte FROM DUAL)"
+    return "$PROMPT AS (SELECT $promptsCte ${if (useDualTable) "FROM DUAL" else ""})"
   }
 
   private fun buildPromptsQueryBasedOnType(prompt: Prompt): String = when (prompt.type) {
@@ -149,12 +350,12 @@ class AthenaApiRepository(
     else -> "'${prompt.value}' AS ${prompt.name}"
   }
 
-  private fun buildContextQuery(userToken: DprAuthAwareAuthenticationToken?): String = """WITH $CONTEXT AS (
+  private fun buildContextQuery(userToken: DprAuthAwareAuthenticationToken?, useDualTable: Boolean = true): String = """WITH $CONTEXT AS (
       SELECT 
       '${userToken?.getUsername()}' AS username, 
       '${userToken?.getActiveCaseLoadId()}' AS caseload, 
       'GENERAL' AS account_type 
-      FROM DUAL
+      ${if (useDualTable) "FROM DUAL" else ""}
       )"""
 
   private fun buildFinalQuery(
@@ -169,16 +370,6 @@ class AthenaApiRepository(
     val query = listOf(context, prompts, datasetQuery, reportQuery, policiesQuery, filtersQuery).joinToString(",") + "\n$selectFromFinalStageQuery"
     log.debug("Database query: $query")
     return query
-  }
-
-  private fun mapAthenaStateToRedshiftState(queryState: String): String {
-    val athenaToRedshiftStateMappings = mapOf(
-      "QUEUED" to "SUBMITTED",
-      "RUNNING" to QUERY_STARTED,
-      "SUCCEEDED" to QUERY_FINISHED,
-      "CANCELLED" to QUERY_ABORTED,
-    )
-    return athenaToRedshiftStateMappings.getOrDefault(queryState, queryState)
   }
 
   private fun calculateDuration(status: QueryExecutionStatus): Long = status.completionDateTime()?.let { completion ->