+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/Template.kt
@@ -20,4 +20,7 @@ enum class Template {
 
   @SerializedName("parent-child")
   ParentChild,
+
+  @SerializedName("parent-child-section")
+  ParentChildSection,
 }