+++ src/test/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/ReportDefinitionMapperTest.kt
@@ -18,6 +18,7 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.F
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.FilterType.Text
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.SingleVariantReportDefinition
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.IdentifiedHelper
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.alert.AlertCategory
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.establishmentsAndWings.EstablishmentToWing
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.establishmentsAndWings.EstablishmentToWing.Companion.ALL_WINGS
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.Dataset
@@ -56,6 +57,7 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policye
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.policyengine.Rule
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.security.DprAuthAwareAuthenticationToken
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.DefinitionMapper.Companion.DEFAULT_MAX_STATIC_OPTIONS
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.alert.AlertCategoryCacheService
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.estcodesandwings.EstablishmentCodesToWingsCacheService
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service.model.Caseload
 import java.math.BigDecimal
@@ -70,8 +72,8 @@ class ReportDefinitionMapperTest {
   private val fullDataset = Dataset(
     id = "10",
     name = "11",
-    query = "12",
     datasource = "12A",
+    query = "12",
     schema = Schema(
       field = listOf(
         SchemaField(
@@ -246,11 +248,12 @@ class ReportDefinitionMapperTest {
   private val authToken = mock<DprAuthAwareAuthenticationToken>()
   private val identifiedHelper = IdentifiedHelper()
   private val establishmentCodesToWingsCacheService = mock<EstablishmentCodesToWingsCacheService>()
+  private val alertCategoryCacheService: AlertCategoryCacheService = mock()
+
+  val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService, alertCategoryCacheService)
 
   @Test
   fun `Getting report for user maps full data correctly`() {
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
     val result = mapper.mapReport(definition = singleReportProductDefinition, userToken = authToken)
 
     assertThat(result).isNotNull
@@ -377,7 +380,6 @@ class ReportDefinitionMapperTest {
       allDatasets = listOf(fullDataset),
       allReports = emptyList(),
     )
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
 
     val result = mapper.mapReport(productDefinition, authToken)
 
@@ -399,7 +401,7 @@ class ReportDefinitionMapperTest {
     val defaultValue = createProductDefinition("today($offset, $magnitude)")
     val expectedDate = getExpectedDate(offset, magnitude)
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].filter!!.defaultValue).isEqualTo(expectedDate)
 
@@ -411,7 +413,7 @@ class ReportDefinitionMapperTest {
     val defaultValue = createProductDefinition("today()")
     val expectedDate = LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE)
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].filter!!.defaultValue).isEqualTo(expectedDate)
 
@@ -426,7 +428,7 @@ class ReportDefinitionMapperTest {
     val expectedDate3 = getExpectedDate(7, ChronoUnit.DAYS)
     val expectedResult = "$expectedDate1, $expectedDate2, $expectedDate3"
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].filter!!.defaultValue).isEqualTo(expectedResult)
 
@@ -452,7 +454,7 @@ class ReportDefinitionMapperTest {
     )
     val expectedDate = getExpectedDate(offset, magnitude)
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].filter!!.min).isEqualTo(expectedDate)
     assertThat(result.variant.specification!!.fields[0].filter!!.max).isEqualTo(expectedDate)
@@ -462,8 +464,6 @@ class ReportDefinitionMapperTest {
 
   @Test
   fun `Getting single report for user maps full data correctly`() {
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
     val result = mapper.mapReport(fullSingleReportProductDefinition, authToken)
 
     assertThat(result).isNotNull
@@ -516,8 +516,6 @@ class ReportDefinitionMapperTest {
 
     val fullSingleProductDefinition = fullSingleReportProductDefinition.copy(report = reportWithDynamicFilter)
 
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
     whenever(
       configuredApiService.validateAndFetchData(
         fullSingleProductDefinition.id,
@@ -566,8 +564,6 @@ class ReportDefinitionMapperTest {
       allDatasets = listOf(establishmentDataset),
     )
 
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
     whenever(
       configuredApiService.validateAndFetchDataForFilterWithDataset(any(), any(), any()),
     ).thenReturn(
@@ -603,8 +599,6 @@ class ReportDefinitionMapperTest {
     val reportWithDynamicFilter = generateReport(DynamicFilterOption(minimumLength = 2, returnAsStaticOptions = false))
     val fullSingleProductDefinition = fullSingleReportProductDefinition.copy(report = reportWithDynamicFilter)
 
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
     whenever(
       configuredApiService.validateAndFetchData(
         fullSingleProductDefinition.id,
@@ -632,8 +626,6 @@ class ReportDefinitionMapperTest {
     val reportWithMakeUrlFormula = createReport("make_url('\${profile_host}/prisoner/\${prisoner_number}',\${full_name},TRUE)")
 
     val fullSingleProductDefinition = fullSingleReportProductDefinition.copy(report = reportWithMakeUrlFormula)
-
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
 
     val result = mapper.mapReport(fullSingleProductDefinition, authToken)
 
@@ -677,8 +669,8 @@ class ReportDefinitionMapperTest {
         reportDataset = Dataset(
           id = "10",
           name = "11",
-          query = "12",
           datasource = "12A",
+          query = "12",
           schema = Schema(
             field = listOf(
               SchemaField(
@@ -691,8 +683,6 @@ class ReportDefinitionMapperTest {
           ),
         ),
       )
-
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
 
     val result = mapper.mapReport(fullSingleProductDefinition, authToken)
 
@@ -738,7 +728,7 @@ class ReportDefinitionMapperTest {
       visible = visibleDpd,
     )
 
-    val result: SingleVariantReportDefinition = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result: SingleVariantReportDefinition = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].visible).isEqualTo(visibleControllerModel)
     assertThat(result.variant.specification!!.fields[0].mandatory).isEqualTo(mandatoryControllerModel)
@@ -761,7 +751,7 @@ class ReportDefinitionMapperTest {
       reportFieldDisplay = reportDisplay,
     )
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].display).isEqualTo(expectedDisplay)
 
@@ -781,8 +771,6 @@ class ReportDefinitionMapperTest {
       mandatory = true,
     )
     val productDefinition = createProductDefinition("today()", parameters = listOf(parameter))
-
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
 
     val result = mapper.mapReport(productDefinition, authToken)
 
@@ -840,10 +828,9 @@ class ReportDefinitionMapperTest {
         ),
       ),
     )
+    val rdfMapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService, alertCategoryCacheService)
 
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
-    val result = mapper.mapReport(productDefinition, authToken)
+    val result = rdfMapper.mapReport(productDefinition, authToken)
 
     val matchingField = result.variant.specification!!.fields.filter { it.name == parameterName }
 
@@ -897,8 +884,6 @@ class ReportDefinitionMapperTest {
       ),
     )
 
-    val mapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService)
-
     val result = mapper.mapReport(productDefinition, authToken)
 
     val matchingField = result.variant.specification!!.fields.filter { it.name == parameterName }
@@ -918,10 +903,57 @@ class ReportDefinitionMapperTest {
   }
 
   @Test
+  fun `getting single report with parameters with referenceType of alert_code includes all the alerts as static options`() {
+    val parameterName = "paramName"
+    val parameterDisplay = "paramDisplay"
+    val parameter = Parameter(
+      index = 0,
+      name = parameterName,
+      reportFieldType = ParameterType.String,
+      filterType = FilterType.Text,
+      display = parameterDisplay,
+      mandatory = true,
+      referenceType = ReferenceType.ALERT,
+    )
+    val productDefinition = createProductDefinition("today()", parameters = listOf(parameter))
+    val alertCode1 = "OPPO"
+    val alertDesc1 = "PPO Case"
+    val alertCode2 = "P1"
+    val alertDesc2 = "MAPPA Level 1 Case"
+
+    whenever(
+      alertCategoryCacheService.getAlertCodesCacheIfNeeded(),
+    ).thenReturn(
+      mapOf(
+        "ALERT_CODES" to listOf(
+          AlertCategory("ALERT_CODE", alertCode1, alertDesc1),
+          AlertCategory("ALERT_CODE", alertCode2, alertDesc2),
+        ),
+      ),
+    )
+    val rdfMapper = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService, alertCategoryCacheService)
+
+    val result = rdfMapper.mapReport(productDefinition, authToken)
+
+    val matchingField = result.variant.specification!!.fields.filter { it.name == parameterName }
+
+    val expectedStaticOptions = listOf(
+      FilterOption(alertCode1, alertDesc1),
+      FilterOption(alertCode2, alertDesc2),
+      FilterOption("All", "All"),
+    )
+
+    val expectedReportField = createReportFieldDefinition(parameter, expectedStaticOptions)
+    assertThat(result.variant.specification!!.fields.size).isEqualTo(2)
+    assertThat(matchingField.size).isEqualTo(1)
+    assertThat(matchingField[0]).isEqualTo(expectedReportField)
+  }
+
+  @Test
   fun `Interactive report metadata hint is mapped to the report correctly`() {
     val defaultValue = createProductDefinition("today(-2,DAYS)", interactive = true)
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.interactive).isEqualTo(true)
   }
@@ -931,8 +963,8 @@ class ReportDefinitionMapperTest {
     val sourceDataset = Dataset(
       id = "10",
       name = "11",
-      query = "12",
       datasource = "12A",
+      query = "12",
       schema = Schema(
         field = listOf(
           SchemaField(
@@ -973,7 +1005,7 @@ class ReportDefinitionMapperTest {
       allReports = listOf(fullReport, childReport),
     )
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = sourceDefinition, userToken = authToken)
+    val result = mapper.mapReport(definition = sourceDefinition, userToken = authToken)
 
     assertThat(result.variant.specification!!.fields[0].filter?.type.toString()).isEqualTo("Text")
     assertThat(result.variant.specification!!.fields[0].filter?.mandatory).isTrue()
@@ -1006,7 +1038,7 @@ class ReportDefinitionMapperTest {
       ),
     )
 
-    val result = ReportDefinitionMapper(configuredApiService, identifiedHelper, establishmentCodesToWingsCacheService).mapReport(definition = defaultValue, userToken = authToken)
+    val result = mapper.mapReport(definition = defaultValue, userToken = authToken)
 
     assertThat(result.variant.specification?.fields?.first()).isEqualTo(
       FieldDefinition(
@@ -1182,8 +1214,8 @@ class ReportDefinitionMapperTest {
     Dataset(
       id = "10",
       name = "11",
-      query = "12",
       datasource = "12A",
+      query = "12",
       schema = Schema(
         field = listOf(
           SchemaField(